Code: Fixes VC warnings

+++ b64/encode.h
@@ -12,7 +12,7 @@ For details, see http://sourceforge.net/projects/libb64
 
 namespace base64
 {
-	extern "C" 
+	extern "C"
 	{
 		#include "cencode.h"
 	}
@@ -54,7 +54,7 @@ namespace base64
 			do
 			{
 				istream_in.read(plaintext, N);
-				plainlength = istream_in.gcount();
+				plainlength = static_cast<int>(istream_in.gcount());
 				//
 				codelength = encode(plaintext, plainlength, code);
 				ostream_in.write(code, codelength);
